add opencv draw

+++ pose_estimate/src/estimator.py
@@ -177,13 +177,10 @@ class Gen6DEstimator:
             # stage 1: detection
             with torch.no_grad():
                 detection_outputs = self.detector.detect_que_imgs(que_img[None])
-                # position = detection_outputs['positions'][0]
-                position = np.array((result.position[0], result.position[1]), dtype=float) 
-                # 288.31536865234375, 351.4462890625
-                # position = np.array((288, 351), dtype=float)
+                position = np.array((int(result.position[0]), int(result.position[1])), dtype=int)
                 scale_r2q = np.array((result.size/75), dtype=float) 
+                # position = detection_outputs['positions'][0]
                 # scale_r2q = detection_outputs['scales'][0]
-                # print(scale_r2q)
 
             # crop the image according to the detected scale and the detected position
             que_img_, _ = transformation_crop(que_img, position, 1/scale_r2q, 0, self.cfg['ref_resolution'])  # h,w,3
